Swap dependency to build-flow-plugin [JENKINS-19470]
Swap the dependency so that buildgraph-view is dependent of
build-flow-plugin instead of build-flow-plugin being optionally
dependent on buildgraph-view. This seems to be the better way to
do it since that is how they actually are dependent on each other
and buildgraph-view can access functions and classes in the
build-flow-plugin that it might need to customize what is shown in
the graph.