[FIXED JENKINS-23607] Do not block thread waiting for builds in queue to be scheduled
The json response handler calls getBuild() on all JobInvocations in the graph.
If a job is not scheduled (still in build queue), the call getBuild() is blocking.
This results in a HTTP request thread being blocked.
If enough such threads gets blocked, jenkins stops responding to HTTP requests all together.
To prevent this scenario this fix only calls getBuild() if the build has been started.